Fix segfault affecting resizing of large streams with small tiles
Concurrent getTileImage() calls for the same tileIndex can occur
unexpectedly when resizing a stream window. Adding a per-tile mutex is
the simplest way to fix the problem and should not penalize performance
too much as those events are rare. Returning an empty ImagePtr does not
work (the stream blocks).

+++ tide/wall/tools/PixelStreamChannelAssembler.cpp
@@ -130,7 +130,7 @@ Indices PixelStreamChannelAssembler::computeVisibleSet(
     return visibleSet;
 }
 
-uint PixelStreamChannelAssembler::getTilesCount() const
+size_t PixelStreamChannelAssembler::getTilesCount() const
 {
     return _getTilesX() * _getTilesY();
 }
